requirements.txt
PyGithub>=2.1.1
pytest>=7.3.1
pylint>=2.17.0
PyYAML>=6.0.1